py-fastmcp-server/pyproject.toml
[project]
name = "py-fastmcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = ["mcp[cli]>=1.10.1", "pillow>=11.3.0"]
